Updated layout and added logic to login
Layout of the login screen now looks better on larger screens, also gives an example of using Angular materials layout options. If a user enters an invalid login they will be notified on the login screen, this shows an example of catching errors from the loopback service and doing something with them.